make wide-SMP code clippy- and rustfmt-clean
The wide-SMP topology/IOAPIC work landed without a clean cargo fmt or
cargo clippy -D warnings, both of which the CI lint job gates (cargo fmt
-- --check; cargo clippy --workspace --all-targets on the default and
wprof,integration legs; cargo doc -D warnings).

cargo fmt reflows 16 files to rustfmt defaults (fn-call args past
fn_call_width one-per-line, method-chain wrapping, assert! arg wrapping,
trailing commas).

clippy -D warnings, 15 errors, all behavior-preserving:
- collapsible_if to let-chains in the IOAPIC MMIO read/write dispatch,
  the perf-mode and max-leaf hypervisor-CPUID gates, and the teardown
  routing-failure summary. The max-leaf gate keeps explicit parens,
  (wide_smp || performance_mode) &&, so the ||/&& precedence is unchanged.
- manual_range_contains to Range::contains for the IOAPIC MMIO window.
- io_other_error to io::Error::other in an IOAPIC-routing test injector.
- doc_lazy_continuation: escape a >N line that read as a stray blockquote.
- unnecessary_cast (a redundant usize cast) and manual_is_multiple_of in
  two topology tests.

+++ tests/sparse_apic_id_e2e.rs
@@ -18,7 +18,7 @@
 //! This pins the invariant on a MINIMAL sparse topology — 2 LLC x 3 core x 1
 //! thread = 6 vCPUs; 3 cores need a 2-bit field, so `apic_id = (llc<<2)|core`
 //! = {0,1,2,4,5,6} (gap at 3, max 6 > 5) — so it runs on any host without a
-//! >255-vCPU box. The >255 ext-dest path itself is exercised by the
+//! \>255-vCPU box. The >255 ext-dest path itself is exercised by the
 //! `wide_smp_*_irq_e2e` tests on a >=252-CPU host; this is the CI-runnable
 //! root-cause guard. `/proc/cpuinfo` exposes both `apicid` (runtime,
 //! `read_apic_id()`) and `initial apicid` (CPUID leaf 1) per CPU; the fix
@@ -71,7 +71,7 @@ fn read_apicids() -> Result<Vec<(u32, u32)>> {
 
 #[ktstr_test(llcs = 2, cores = 3, threads = 1, no_perf_mode, duration_s = 4)]
 fn sparse_topology_runtime_apicid_matches_declared(ctx: &Ctx) -> Result<AssertResult> {
-    let total = ctx.topo.total_cpus() as usize;
+    let total = ctx.topo.total_cpus();
     ensure!(
         total == 6,
         "expected 6 vCPUs (2 LLC x 3 core x 1 thread), got {total}"
